app/requirements.txt
streamlit==1.48.1
pillow==10.4.0 
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.5.1
fastapi==0.115.11